Update Go version and add main function in for_loop
Updated the Go version from 1.21 to 1.22 to ensure compatibility with the latest features and fix potential bugs related to the old version. Also, a new main function was included within the 'for_loop' directory to add more functionality to the project, featuring a concurrent execution structure using goroutines, synchronisation and sleep timers.